+++ src/main/java/ru/saynurdinov/moviefan/repository/DirectorRepository.java
@@ -1,7 +1,11 @@
 package ru.saynurdinov.moviefan.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import ru.saynurdinov.moviefan.model.Actor;
 import ru.saynurdinov.moviefan.model.Director;
 
-public interface DirectorRepository extends JpaRepository<Director, Integer> {
+import java.util.List;
+
+public interface DirectorRepository extends JpaRepository<Director, Long> {
+
 }
